# docs/reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: Narad HTTP API
  version: v1
  summary: Produce, consume and manage topics on a Narad cluster over HTTP.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  x-narad-lead: |
    Look up every Narad HTTP endpoint: its parameters, request body, the
    grant it needs, the status codes it returns and a real example.
  x-narad-lead-example:
    request: |
      curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
    response: |
      HTTP/1.1 200 OK
      Content-Length: 456
      Content-Type: application/json
      Date: Mon, 28 Sep 2026 19:31:43 GMT

      {
        "name": "orders",
        "id": "681a429c0b683b2d",
        "partitions": 3,
        "retention_ms": 172800000,
        "visibility_timeout_ms": 30000,
        "max_in_flight_per_partition": 1024,
        "max_acked_ahead_per_partition": 1024,
        "created_at": 1790623903,
        "owner": "admin",
        "role": "parent",
        "children": ["orders-audit"],
        "schema_version": 0,
        "partition_stats": [
          {
            "index": 1,
            "segments": 1,
            "oldest_offset": 0,
            "next_offset": 2,
            "high_watermark": 2,
            "size_bytes": 166,
            "oldest_segment_at": 1790623903,
            "owner_node": "narad-0"
          }
        ]
      }
  x-narad-placeholders: |
    - `$NARAD` is the base URL of any node or of the load balancer in front
...
      off and `-u "$AUTH"` can be left out.

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /v1/topics/{topic}/produce:
    parameters:
      - $ref: '#/components/parameters/TopicPath'
    post:
      operationId: produce
      tags: [Messages]
      summary: Produce a message
      description: |
        Stores the request body as one message. A
        [`202`](status-codes.md#status-202) means the node that answered has
        written the message to its [ingress WAL](glossary.md#ingress-wal)
        and synced it to disk; it then moves the message to the partition's
        owner in the background. What a `202` promises is in the
        [delivery contract](../understand/delivery-contract.md#what-202-means).

        The body is stored byte for byte. Send JSON with
        `Content-Type: application/json` and anything else with
        `Content-Type: application/octet-stream`. On a topic with a schema,
        the body must be one JSON text that the current schema accepts
        ([Schema validation rules](schema-rules.md#validation)).
      x-narad-grant: |
        `produce` on the topic.
      parameters:
        - name: key
          in: query
          description: |
            Routes the message: messages with the same key go to the same
            partition while the partition count is unchanged and its owner
            is up. This is not an ordering guarantee. Without a key, messages
            are spread round-robin.
          schema:
            type: string
        - name: partition
          in: query
          description: Pins the message to this partition. Wins over `key`.
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        description: The message, 1 byte to 1 MiB.
        content:
          application/json:
            schema: {}
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Accepted and synced to disk. The body is empty.
        '400':
          description: Empty body, `partition` out of range, `key` or `partition` given twice, or the body fails the topic's schema (from v3.1.0, that includes a body nested deeper than 256 levels).
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No `produce` grant on the topic.
        '404':
          description: The topic does not exist.
        '409':
          description: 'The topic is a delay child, which only its parent can feed. Unreleased: the topic is a remote child''s stub, whose messages live on its remote (`remote child "<name>" lives on remote <remote>; consume it there`).'
        '413':
          description: The body is over 1 MiB.
        '415':
          description: No accepted `Content-Type` and no `X-Narad-Client` header.
        '429':
          description: Too many produces in flight for this user on this node, only when the operator set a produce cap ([Configuration reference](configuration.md#http)).
        '500':
          description: The node could not write to its ingress WAL. It answers every produce this way until it restarts.
        '503':
          description: 'New in v3.1.0. Nothing was checked or stored; retry through another node. Either this node is being decommissioned and takes no new produce (with `Retry-After: 1`), or the topic has a schema and every schema validation slot on the node stayed busy for 5 seconds.'
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X POST \
              "$NARAD/v1/topics/orders/produce?key=customer-42" \
              -H "Content-Type: application/json" \
              -d '{"order_id": "ord_123", "amount": 1250}'
          response: |
            HTTP/1.1 202 Accepted
            Date: Mon, 28 Sep 2026 19:31:43 GMT
            Content-Length: 0
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    TopicPath:
      name: topic
      in: path
      required: true
      description: Name of the topic.
      schema:
        type: string
...